refactor: break down styles and move preview to component
