+++ packages/core/components/LayerTree/components/drop-line/styles.module.css
@@ -1,0 +1,26 @@
+/* Private --_puck-outline-* vars resolve on .LayerTreeRoot. */
+.DropLine {
+  background: var(--_puck-outline-color-drop-indicator);
+  border-radius: calc(var(--_puck-outline-drop-indicator-size) / 2);
+  height: var(--_puck-outline-drop-indicator-size);
+  inset-inline: 0;
+  pointer-events: none;
+  position: absolute;
+  z-index: 1;
+}
+
+.DropLine--top {
+  top: 0;
+}
+
+.DropLine--bottom {
+  bottom: 0;
+}
+
+.DropLine--top.DropLine--outset {
+  top: calc(-1 * var(--_puck-outline-drop-indicator-size));
+}
+
+.DropLine--bottom.DropLine--outset {
+  bottom: calc(-1 * var(--_puck-outline-drop-indicator-size));
+}
