+++ packages/core/components/LayerTree/components/layer-tree-items/index.ts
@@ -1,0 +1,2 @@
+export { StaticLayerTreeItems } from "./static-layer-tree-items";
+export { VirtualizedLayerTreeItems } from "./virtualized-layer-tree-items";
