+++ packages/core/components/LayerTree/components/layer-tree-items/styles.module.css
@@ -1,0 +1,17 @@
+/* Private --_puck-outline-* vars resolve on .LayerTreeRoot. */
+.LayerTree {
+  color: var(--_puck-outline-color-text);
+  font-family: var(
+    --puck-outline-font-family,
+    var(--puck-font-family-monospaced)
+  );
+  font-size: var(--puck-outline-font-size, var(--puck-font-size-xxxs));
+  margin: 0;
+  position: relative;
+  list-style: none;
+  padding: 0;
+}
+
+.LayerTree--nested {
+  margin-inline-start: var(--puck-outline-space-indent, var(--puck-space-3));
+}
